Update: registrations account to mutable
Previously, the account was not writatble, within the RegisterCandidate
account context. This meant that it was impossible to update the value
within the register_candidate instruction handler. Subsequently, it
would result in a seed constraint violation for the candidate account in the integrated tests, since the registrations aaccount `count` attribute is used as a seed for the candidate pda